github-pr-autofix/backend/requirements.txt
Flask
Flask-CORS
requests
python-dotenv
Werkzeug